requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.10.0
tqdm>=4.65.0
pandas>=1.5.0
scikit-learn>=1.2.0
flask>=2.3.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.28.0
gunicorn>=20.1.0
nltk>=3.8.1 